Imporve some main screen visuality
